+++ extensions/omni_id/src/identity_type.sql
@@ -8,7 +8,7 @@ create function identity_type(name name, type regtype default 'int8',
                               constructor text default null,
                               create_constructor bool default true,
                               operator_schema name default 'public',
-                              nextval regproc default null
+                              nextval regprocedure default null
 ) returns regtype
     language plpgsql
 as
@@ -185,8 +185,8 @@ begin
     -- If sequence API is desired for UUID, supply it as `_nextval`
     if not create_sequence and nextval is not null then
         execute format(
-                'create function %I() returns %I language sql as $sql$ select %I()::%4$I.%2$I $sql$',
-                name || '_nextval', name, nextval, ns);
+                'create function %I() returns %I language sql as $sql$ select %s()::%4$I.%2$I $sql$',
+                name || '_nextval', name, nextval::regproc, ns);
     end if;