+++ extensions/omni/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.5] - TBD
+
+### Added
+
+* Support for the upcoming Postgres 18 [#763](https://github.com/omnigres/omnigres/pull/763)
+
 ## [0.2.4] - 2024-12-06
 
 ### Fixed
@@ -113,3 +119,5 @@ Initial release following a few months of iterative development.
 [0.2.3]: [https://github.com/omnigres/omnigres/pull/699]
 
 [0.2.4]: [https://github.com/omnigres/omnigres/pull/696]
+
+[0.2.5]: [https://github.com/omnigres/omnigres/pull/763]