+++ extensions/omni_id/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - TBD
+
+### Added
+
+* Support for unreleased Postgres versions [#763](https://github.com/omnigres/omnigres/pull/763)
+
 ## [0.4.0] - 2024-12-16
 
 ### Removed
@@ -61,3 +67,5 @@ Initial release
 [0.3.1]: [https://github.com/omnigres/omnigres/pull/644]
 
 [0.4.0]: [https://github.com/omnigres/omnigres/pull/707]
+
+[0.4.1]: [https://github.com/omnigres/omnigres/pull/763]