+++ extensions/omni_sql/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - TBD
+
+### Added
+
+* Support for the upcoming Postgres 18 [#763](https://github.com/omnigres/omnigres/pull/763)
+
 ## [0.5.0] - 2025-01-20
 
 ### Added
@@ -126,3 +132,9 @@ Initial release following a few months of iterative development.
 [0.3.6]: [https://github.com/omnigres/omnigres/pull/717]
 
 [0.3.7]: [https://github.com/omnigres/omnigres/pull/718]
+
+[0.4.0]: [https://github.com/omnigres/omnigres/pull/747]
+
+[0.5.0]: [https://github.com/omnigres/omnigres/pull/753]
+
+[0.5.1]: [https://github.com/omnigres/omnigres/pull/763]